WIP: Tightening Suricata Bolts
Added logic to disallow editing of community rules. Adjusted detection's form validation rules so that community rules aren't validated at the form level.

Removed "Details" Detection tab and reordered the controls under it to the "Summary" and "Signature" tabs. Rearranged some tabs.

Cleaned up the strings used in dropdowns (engine, severity) with capitalization. Cleaner UI, but the correct casing is still passed over the wire.

Refactored PublicID's "Generate" button to be "Extract" so that it fills the field using the SID in the rule. May eventually make this field readonly except for this modification.

Updated infohandler's response to use the same severity values the rules use.

Converted more strings to i18n references.

`indexDocument` and `deleteDocument` on the ElasticEventstore now pass the ctx in. This removes the WARN log message about making a request without a user. Needs more thorough testing to be sure there's no unwanted side effects.